计算订单金额

+++ toucan-shopping-modules/order/toucan-shopping-order-entity-vo/src/main/java/com/toucan/shopping/modules/order/entity/Order.java
@@ -2,6 +2,7 @@ package com.toucan.shopping.modules.order.entity;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 /**
@@ -14,9 +15,9 @@ public class Order {
     private Long id; //主键
     private String orderNo; //订单编号
     private String userId; //用户ID
-    private Double orderAmount; //订单金额
-    private Double payAmount; //付款金额
-    private Double totalAmount; //商品最终金额(折扣算完)
+    private BigDecimal orderAmount; //订单金额
+    private BigDecimal payAmount; //付款金额
+    private BigDecimal totalAmount; //商品最终金额(折扣算完)
     private Integer payStatus; //支付状态 0未支付 1已支付 3线下支付已到账 4取消支付
     private Integer tradeStatus; //交易状态 0进行中 1已完成 2已取消交易 3已结算
     private Integer payMethod;  //支付方式 1线上支付 2线下支付
